+++ CHANGELOG.md
@@ -63,3 +63,7 @@ sorting for any reason.
 
 ## 0.0.5
 Refactored the merge sort method to use recursive nodes instead of a map. Also added a stream for the maximum number of choices left if the user sorts inefficiently.
+
+## 0.0.6
+Added a 'partiallySorted' constructor that allows you to add a list of unsorted items and a list of sorted lists. 
+The contents of a sorted list will not be compared against each other.